move iter_snapshots to storage/db
