+++ crates/primitives/src/snapshot/mod.rs
@@ -4,17 +4,11 @@ mod compression;
 mod filters;
 mod segment;
 
-use alloy_primitives::{BlockNumber, TxNumber};
+use alloy_primitives::{BlockNumber};
 pub use compression::Compression;
 pub use filters::{Filters, InclusionFilter, PerfectHashingFunction};
 pub use segment::{SegmentConfig, SegmentHeader, SnapshotSegment};
 
-use crate::fs::FsPathError;
-use std::{
-    collections::{hash_map::Entry, HashMap},
-    ops::RangeInclusive,
-    path::Path,
-};
 
 /// Default snapshot block count.
 pub const BLOCKS_PER_SNAPSHOT: u64 = 500_000;
@@ -51,40 +45,4 @@ impl HighestSnapshots {
             SnapshotSegment::Receipts => &mut self.receipts,
         }
     }
-}
-
-/// Alias type for a map of [`SnapshotSegment`] and sorted lists of existing snapshot ranges.
-type SortedSnapshots =
-    HashMap<SnapshotSegment, Vec<(RangeInclusive<BlockNumber>, RangeInclusive<TxNumber>)>>;
-
-/// Given the snapshot's location, it returns a list over the existing snapshots organized by
-/// [`SnapshotSegment`]. Each segment has a sorted list of block ranges and transaction ranges.
-pub fn iter_snapshots(path: impl AsRef<Path>) -> Result<SortedSnapshots, FsPathError> {
-    let mut static_files: HashMap<SnapshotSegment, Vec<_>> = HashMap::default();
-    let entries = crate::fs::read_dir(path.as_ref())?.filter_map(Result::ok).collect::<Vec<_>>();
-
-    for entry in entries {
-        if entry.metadata().map_or(false, |metadata| metadata.is_file()) {
-            if let Some((segment, block_range, tx_range)) =
-                SnapshotSegment::parse_filename(&entry.file_name())
-            {
-                let ranges = (block_range, tx_range);
-                match static_files.entry(segment) {
-                    Entry::Occupied(mut entry) => {
-                        entry.get_mut().push(ranges);
-                    }
-                    Entry::Vacant(entry) => {
-                        entry.insert(vec![ranges]);
-                    }
-                }
-            }
-        }
-    }
-
-    // Sort by block end range.
-    for (_, range_list) in static_files.iter_mut() {
-        range_list.sort_by(|a, b| a.0.end().cmp(b.0.end()));
-    }
-
-    Ok(static_files)
 }
